[ADD] API Gateway เพิ่มตัว Filter

+++ Microservice/APIGateway/src/main/java/com/meetu/gateway/configuration/FirebaseConfiguration.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package meetu.userservice.config;
+package com.meetu.gateway.configuration;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
@@ -13,7 +13,6 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import meetu.userservice.controller.UserController;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -27,8 +26,11 @@ public class FirebaseConfiguration {
     @Bean
     public void initFirebaseApp() {
         FileInputStream serviceAccount = null;
-        FirebaseApp.getInstance().delete();
         try {
+            if (FirebaseApp.DEFAULT_APP_NAME == null) {
+                System.out.println("Destroy Firebase Inastance");
+                FirebaseApp.getInstance().delete();
+            }
             System.out.println("---------- Firebase Configuration Start ---------------");
             System.out.println(FirebaseApp.DEFAULT_APP_NAME);
             serviceAccount = new FileInputStream("C:\\ProjectCode\\MeetU\\Microservice\\UserService\\meetu-firebase-key.json");
